Try using ExtraVolumes for file customization
We'd like to let users customize some config files in telemetry.
Right now, we provide some default contents for these files, which
are managed by kolla, but we have requests to allow customization.
What more, CloudKitty requires such a feature to even be usable.

Issue is, we need to keep providing the default files as we've done
until now. But if an ExtraVolume is defined to replace a file, which
is then replaced by kolla, kolla gets permission denied and the
container start fails. So if we were to use ExtraVolumes, we'd need
to filter out the default files being replaced, so that kolla doesn't
try to replace them. This is a working demonstration of this approach.